refactor(jsdoc-cli): move flag-parsing code to flags.js
